fix: The share page query is invalid
