+++ ui/src/pages/DataLogs/components/RawLogsIndexes/IndexHeader/index.less
@@ -6,7 +6,7 @@
   padding: 5px 5px 0 5px;
   .title {
     font-weight: bold;
-    font-size: 16px;
+    font-size: 14px;
     user-select: none;
   }
   .icon {
