+++ ui/src/pages/DataLogs/components/RawLogsIndexes/IndexHeader/index.tsx
@@ -1,20 +1,20 @@
-import indexHeaderStyles from '@/pages/DataLogs/components/RawLogsIndexes/IndexHeader/index.less';
-import IconFont from '@/components/IconFont';
-import { Tooltip } from 'antd';
-import { useModel } from '@@/plugin-model/useModel';
+import indexHeaderStyles from "@/pages/DataLogs/components/RawLogsIndexes/IndexHeader/index.less";
+import IconFont from "@/components/IconFont";
+import { Tooltip } from "antd";
+import { useModel } from "@@/plugin-model/useModel";
 type IndexHeaderProps = {};
 const IndexHeader = (props: IndexHeaderProps) => {
-  const { onChangeVisibleIndexModal } = useModel('dataLogs');
+  const { onChangeVisibleIndexModal } = useModel("dataLogs");
   return (
     <div className={indexHeaderStyles.indexHeaderMain}>
-      <span className={indexHeaderStyles.title}>索引列表</span>
+      <span className={indexHeaderStyles.title}>分析</span>
       <div className={indexHeaderStyles.icon}>
-        <Tooltip title={'索引管理'}>
+        <Tooltip title={"索引管理"}>
           <IconFont
             onClick={() => {
               onChangeVisibleIndexModal(true);
             }}
-            type={'icon-index'}
+            type={"icon-index"}
           />
         </Tooltip>
       </div>
