+++ ui/src/pages/DataLogs/components/RawLogsIndexes/IndexItem/index.less
@@ -2,7 +2,7 @@
   border-bottom: 1px solid hsla(0, 0%, 0%, 0.1);
   .detailContextMain {
     flex: 1;
-    padding: 10px;
+    padding: 10px 0 10px 10px;
     background: #f5f5f5;
     .context {
       display: flex;
@@ -14,7 +14,7 @@
         cursor: default;
         .name {
           display: block;
-          width: 50%;
+          width: 70%;
           overflow: hidden;
           white-space: nowrap;
           text-overflow: ellipsis;
