+++ ui/src/pages/DataLogs/components/RawLogsIndexes/IndexSearchBar/index.tsx
@@ -1,7 +1,7 @@
-import indexSearchBarStyles from '@/pages/DataLogs/components/RawLogsIndexes/IndexSearchBar/index.less';
-import classNames from 'classnames';
-import { Input } from 'antd';
-import { useState } from 'react';
+import indexSearchBarStyles from "@/pages/DataLogs/components/RawLogsIndexes/IndexSearchBar/index.less";
+import classNames from "classnames";
+import { Input } from "antd";
+import { useState } from "react";
 type IndexSearchBarProps = {
   onSearch: (val: string) => void;
 };
@@ -12,7 +12,7 @@ const IndexSearchBar = (props: IndexSearchBarProps) => {
     <div className={classNames(indexSearchBarStyles.indexSearchBarMain)}>
       <Input.Search
         value={value}
-        placeholder={'搜索索引'}
+        placeholder={"搜索索引"}
         allowClear
         onSearch={onSearch}
         onChange={(ev) => setValue(ev.target.value)}
